requirements.txt
sentence-transformers==2.2.2
faiss-cpu==1.13.2
numpy
